Refactored code wrapper to handle variable pipeline states
